feat: setup tsc
