+++ .gitignore
@@ -2,3 +2,4 @@
 *.DS_Store
 **/node_modules
 .idea
+reader/ts/src/*.js
